# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ClinicFlow API Documentation",
        "description": "Comprehensive API documentation for ClinicFlow Backend",
        "contact": {
            "email": "admin@clinicflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "https://api.clinicflow.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/dashboard/inventory": {
            "get": {
                "tags": [
                    "Dashboard - Inventory"
                ],
                "summary": "Get inventory items",
                "operationId": "57f4029612a728c0b568fb4266f80e17",
                "parameters": [
                    {
                        "name": "clinic_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dashboard - Inventory"
                ],
                "summary": "Create inventory item",
                "operationId": "84374918edab0c19a1e1d74f1768de89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clinic_id",
                                    "item_name",
                                    "quantity",
                                    "unit_price",
                                    "reorder_level"
                                ],
                                "properties": {
                                    "clinic_id": {
                                        "type": "integer"
                                    },
                                    "item_name": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "unit_price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/invoices": {
            "get": {
                "tags": [
                    "Dashboard - Invoices"
                ],
                "summary": "Get invoices list",
                "operationId": "dd818e150c7d076daf6979a1bec293b5",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoices list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dashboard - Invoices"
                ],
                "summary": "Create invoice",
                "operationId": "58d68901eb0db5ff7c62b5c7510faa73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "total_amount",
                                    "paid_amount",
                                    "status"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer"
                                    },
                                    "total_amount": {
                                        "type": "number"
                                    },
                                    "paid_amount": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}